backend/requirements.txt
fastapi
uvicorn
python-multipart

pdfplumber
python-docx

sentence-transformers
scikit-learn
python-pptx
numpy